+++ Eurofurence/Application/Shared/Views/Dissolving Title/DissolvingTitleContext.swift
@@ -1,0 +1,8 @@
+import UIKit
+
+protocol DissolvingTitleContext {
+    
+    var title: String? { get }
+    var contextualViewFrame: CGRect { get }
+    
+}
