+++ Eurofurence/Application/Shared/Views/Dissolving Title/DissolvingTitleController.swift
@@ -1,26 +1,5 @@
 import UIKit
 
-protocol DissolvingTitleContext {
-    
-    var title: String? { get }
-    var contextualViewFrame: CGRect { get }
-    
-}
-
-struct DissolvingTitleLabelContext: DissolvingTitleContext {
-    
-    let label: UILabel
-    
-    var title: String? {
-        label.text
-    }
-    
-    var contextualViewFrame: CGRect {
-        label.convert(label.frame, to: nil)
-    }
-    
-}
-
 struct DissolvingTitleController {
     
     private enum TitleState {
