+++ Eurofurence/Application/Shared/Views/Dissolving Title/DissolvingTitleLabelContext.swift
@@ -1,0 +1,15 @@
+import UIKit
+
+struct DissolvingTitleLabelContext: DissolvingTitleContext {
+    
+    let label: UILabel
+    
+    var title: String? {
+        label.text
+    }
+    
+    var contextualViewFrame: CGRect {
+        label.convert(label.frame, to: nil)
+    }
+    
+}
